src/test/java/com/hello/hellospirng/repository/MemoryMemberRepositoryTest.java
package com.hello.hellospirng.repository;

import com.hello.hellospirng.domain.Member;
import org.junit.jupiter.api.AfterEach;
import org.junit.jupiter.api.Test;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

public class MemoryMemberRepositoryTest {

    MemoryMemberRepository repository = new MemoryMemberRepository();

    //테스트 하나가 끝날때마다 깔끔하게 Clean해야함
    @AfterEach
    public void afterEach() {
        repository.clearStore();
    }

    @Test
    public void save() {
        Member member = new Member();
        member.setName("spring");

        repository.save(member);
        //Optional에서는 get()으로 바로 값을 꺼낼 수 있다 하지마 좋은 방법은 아니다.
        Member result = repository.findById(member.getId()).get();
        //assertEquals(기대하는값, 입력값);
        //Assertions.assertEquals(member, result);
        assertThat(member).isEqualTo(result);

    }

    @Test
    public void findByName() {
        Member member1 = new Member();
        member1.setName("spring1");
        repository.save(member1);

        //Rename단축키 shift + F6
        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        Member result = repository.findByName("spring1").get();

        assertThat(result).isEqualTo(member1);
    }

    @Test
    public void findAll() {
        Member member = new Member();
        member.setName("spring1");
        repository.save(member);


        Member member2 = new Member();
        member2.setName("spring2");
        repository.save(member2);

        List<Member> result = repository.findAll();

        assertThat(result.size()).isEqualTo(2);

    }

}
